+++ serialization/base64/src/commonMain/kotlin/dev/inmo/micro_utils/serialization/base64/Base64Serializer.kt
@@ -1,9 +1,6 @@
 package dev.inmo.micro_utils.serialization.base64
 
-import dev.inmo.micro_utils.crypto.*
 import kotlinx.serialization.KSerializer
-import kotlinx.serialization.Serializer
-import kotlinx.serialization.builtins.serializer
 import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
@@ -12,9 +9,9 @@ open class Base64Serializer<T>(
     private val converterFrom: (T) -> String,
     private val converterTo: (String) -> T,
 ) : KSerializer<T> {
-    override val descriptor: SerialDescriptor = String.serializer().descriptor
-    override fun deserialize(decoder: Decoder): T = converterTo(decoder.decodeString().decodeBase64String())
-    override fun serialize(encoder: Encoder, value: T) = encoder.encodeString(converterFrom(value).encodeBase64String())
+    override val descriptor: SerialDescriptor = Base64BytesToFromStringSerializer.descriptor
+    override fun deserialize(decoder: Decoder): T = converterTo(Base64BytesToFromStringSerializer.deserialize(decoder).decodeToString())
+    override fun serialize(encoder: Encoder, value: T) = Base64BytesToFromStringSerializer.serialize(encoder, converterFrom(value).encodeToByteArray())
 }
 
 object Base64StringSerializer : Base64Serializer<String>({ it }, { it })